Add _debug test and share test code
