+++ tests/test_utils.py
@@ -1,4 +1,24 @@
-from protowhat.utils import legacy_signature
+import pytest
+
+from tests.helper import Success, state, dummy_checks
+from protowhat.Test import TestFail as TF
+from protowhat.sct_syntax import ExGen, F
+from protowhat.utils import legacy_signature, _debug
+
+state = pytest.fixture(state)
+dummy_checks = pytest.fixture(dummy_checks)
+
+
+def test_debug(state, dummy_checks):
+    state.do_test(Success("msg"))
+    Ex = ExGen(state, dummy_checks)
+    try:
+        Ex().noop().child_state() >> F(attr_scts={"_debug": _debug})._debug("breakpoint name")
+    except TF as e:
+        assert "breakpoint name" in str(e)
+        assert "history" in str(e)
+        assert "child_state" in str(e)
+        assert "test" in str(e)
 
 
 def test_legacy_signature():
